Mock-test-Agent/requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
langchain==0.1.17
langchain-core==0.1.52
langchain-google-genai==1.0.5
google-generativeai==0.7.2
pydantic==2.6.4
typing_extensions==4.10.0
